Login & Register & File removal
I have removed a ton of useless files that are in here from authentication base, no clue why I even wanted to use it in the first place tbh. It kinda sucks, p.s. volt sucks.

Users can now register & login to their accounts. It's all in-place but I do want to add an email for verification and so on.

I also modified the user table to make it username rather than name.