Flash messages, Table organisation, Routes
Added flash messages for when a user logs in or logs out. This will also be used in the future. Potentially will change this UI anyway.

Table organisation as it was in the wrong directory as I put panels and not panel.

Routes have been organised by creating a RoutesController for the public routes.